feat(liveslots-tools): prepare-strict-test-env
