+++ packages/async-flow/package.json
@@ -37,10 +37,9 @@
     "@endo/promise-kit": "^1.1.5"
   },
   "devDependencies": {
-    "@agoric/swingset-liveslots": "^0.10.2",
+    "@agoric/swingset-vat": "^0.32.2",
     "@agoric/zone": "^0.2.2",
     "@endo/env-options": "^1.1.6",
-    "@endo/ses-ava": "^1.2.5",
     "ava": "^5.3.0",
     "tsd": "^0.31.1"
   },
