+++ packages/zone/package.json
@@ -34,7 +34,7 @@
     "@endo/pass-style": "^1.4.3"
   },
   "devDependencies": {
-    "@agoric/swingset-liveslots": "^0.10.2",
+    "@agoric/swingset-vat": "^0.32.2",
     "@endo/patterns": "^1.4.3",
     "ava": "^5.3.0"
   },
